refactor move store.ts into new store/ directory
